requirements.txt
cython<0.30
numpy<1.17
Pillow<=6.0
